backend/requirements.txt
Flask
Flask-Cors
eventregistry
python-eventregistry
python-dotenv
os
json